third_party/blink/web_tests/platform/generic/dom/legacy_dom_conformance/html/level2/html/AppletsCollection-expected.txt

This test demonstrates a couple of object tags being added that do not contain Java applets. Previously these Object element would be included in the document.applets collection even though the spec says that collection should contain only Applet elements and Object elements that contain applets.

As the document is parsed, javascript will output the size of the document.applets collection after each Object tag is inserted into the document. For this test the count should always be zero.

Tragically we cannot have any actual Java Applets referenced here because when DRT loads the Java VM it hangs. If that bug is resolved this test should also be amended to add a dummy Applet using both the Applet tag and a few different uses of the Object tag.

Applets length is currently 0

Applets length is currently 0

Nothing here!

Applets length is currently 0

Applets length is currently 0
